fix bad code to allow glassfish to deploy the application

+++ src/main/java/ch/heigvd/gamification/services/to/AppUsersTOService.java
@@ -44,7 +44,7 @@ public class AppUsersTOService implements IAppUsersTOService {
   }
 
   @Override
-  public RankedAppUserTO buildRankedUserTO(AppUser source, Long points) {
+  public RankedAppUserTO buildRankedUserTO(AppUser source, Integer points) {
     return new RankedAppUserTO(
             source.getId(),
             points,
